code: Refac blockchain pkg: Sep block and chain
