--- blockchain/blockchain.go
@@ -1,77 +1,0 @@
-package blockchain
-
-import (
-	"crypto/sha256"
-	"errors"
-	"fmt"
-	"sync"
-)
-
-type Block struct {
-	Data     string `json:"data"`
-	Hash     string `json:"hash"`
-	PrevHash string `json:"prevHash,omitempty"`
-	Height   int    `json:"height"`
-}
-
-type blockchain struct {
-	blocks []*Block
-}
-
-var b *blockchain
-var once sync.Once
-
-func (b *Block) calculateHash() {
-	hash := sha256.Sum256([]byte(b.Data + b.PrevHash))
-	b.Hash = fmt.Sprintf("%x", hash)
-}
-
-func (b Block) ShowBlock() (data, hash, prevHash string) {
-	return b.Data, b.Hash, b.PrevHash
-}
-
-func getLastHash() string {
-	totalBlocks := len(GetBlockChain().blocks)
-	if totalBlocks == 0 {
-		return ""
-	}
-	return GetBlockChain().blocks[totalBlocks-1].Hash
-}
-
-func createBlock(data string) *Block {
-	newBlock := Block{data, "", getLastHash(), len(GetBlockChain().blocks) + 1}
-	newBlock.calculateHash()
-	return &newBlock
-}
-
-func (b *blockchain) AddBlock(data string) {
-	b.blocks = append(b.blocks, createBlock(data))
-}
-
-var ErrNotFound = errors.New("block not found")
-
-func (b *blockchain) GetBlock(height int) (*Block, error) {
-	if height > len(b.blocks) {
-		return nil, ErrNotFound
-	}
-	return b.blocks[height-1], nil
-}
-
-func GetBlockChain() *blockchain {
-	if b == nil {
-		// GetBlockChain을 통해 b 인스턴스가 실행되는 상황을 가정해보자.
-		// 그리고 수 많은 go routine 들이 GetBlockChain을 호출한다고 할 때,
-		// b 인스턴스는 한 번만 제대로 초기화 되면 된다.
-		// 그럴 때 동기 처리를 도와주는 sync 패키지를 활용한다.
-		// once.Do()는 메서드 인자에 들어온 함수를 딱 한 번만 실행하는 기능이다.
-		once.Do(func() {
-			b = &blockchain{}
-			b.AddBlock("Genesis Block")
-		})
-	}
-	return b
-}
-
-func (b *blockchain) AllBlocks() []*Block {
-	return b.blocks
-}
